requirements.txt
asgiref
Django
gunicorn
packaging
psycopg2-binary
python-dotenv
sqlparse
tzdata
whitenoise
django-role-permissions
